Tests/LibWeb/Text/expected/Editing/contenteditable-paste-placeholder.txt

paste: GivGivee it a try! <img> @ 7
undo: Give it a try! <img><br> @ 3
redo: GivGivee it a try! <img> @ 7
